SpeedMapping — Module

SpeedMapping

SpeedMapping solves three types of problems:

Accelerating convergent mapping iterations

Solving non-linear systems of equations

Minimizing a function, possibly with box constraints

using two algorithms:

Alternating cyclic extrapolations (ACX)

Anderson Acceleration (AA)

It provides access to recently developed algorithms which, based on benchmarks, are competitive with Julia packages solving similar problems.